requirements.txt
git+http://github.com/django/django.git@stable/1.5.x#egg=django
South==0.7.6
psycopg2
django-registration
django-nose
